+++ packages/gasket-cli/src/commands/start.js
@@ -3,11 +3,6 @@ const GasketCommand = require('../command');
 class StartCommand extends GasketCommand {
   async runHooks() {
     await this.gasket.exec('preboot');
-
-    this.gasket.config = await this.gasket.execWaterfall(
-      'configure',
-      this.gasket.config);
-
     await this.gasket.exec('start');
   }
 }